Improve Modularity
* Create spring-security-kerberos-web to limit dependencies for non-web based projects
* Add org.springframework.security.kerberos.authentication to better align with Spring Security packaging
* Add org.springframework.security.kerberos.authentication.sun to isolate Sun implementations of APIs
* Remove unnecessary dependencies

Fixes gh-20
