+++ settings.gradle
@@ -3,6 +3,7 @@ rootProject.name = 'spring-security-kerberos'
 include 'spring-security-kerberos-core'
 include 'spring-security-kerberos-client'
 include 'spring-security-kerberos-test'
+include 'spring-security-kerberos-web'
 include 'spring-security-kerberos-samples'
 include 'spring-security-kerberos-samples:sec-server-client-auth'
 include 'spring-security-kerberos-samples:sec-server-spnego-form-auth'
