--- spring-security-kerberos-core/src/test/java/org/springframework/security/kerberos/docs/SpnegoConfig.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- tag::snippetA[] -->
-<beans xmlns="http://www.springframework.org/schema/beans"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns:sec="http://www.springframework.org/schema/security"
-  xmlns:context="http://www.springframework.org/schema/context"
-  xsi:schemaLocation="http://www.springframework.org/schema/security http://www.springframework.org/schema/security/spring-security-3.2.xsd
-    http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.1.xsd
-    http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.1.xsd">
-
-  <sec:http entry-point-ref="spnegoEntryPoint" use-expressions="true" >
-    <sec:intercept-url pattern="/" access="permitAll" />
-    <sec:intercept-url pattern="/home" access="permitAll" />
-    <sec:intercept-url pattern="/login" access="permitAll" />
-    <sec:intercept-url pattern="/**" access="authenticated"/>
-    <sec:form-login login-page="/login" />
-    <sec:custom-filter ref="spnegoAuthenticationProcessingFilter"
-      before="BASIC_AUTH_FILTER" />
-  </sec:http>
-
-  <sec:authentication-manager alias="authenticationManager">
-    <sec:authentication-provider ref="kerberosAuthenticationProvider" />
-    <sec:authentication-provider ref="kerberosServiceAuthenticationProvider" />
-  </sec:authentication-manager>
-
-  <bean id="kerberosAuthenticationProvider"
-    class="org.springframework.security.kerberos.KerberosAuthenticationProvider">
-    <property name="userDetailsService" ref="dummyUserDetailsService"/>
-    <property name="kerberosClient">
-      <bean class="org.springframework.security.kerberos.SunJaasKerberosClient">
-        <property name="debug" value="true"/>
-      </bean>
-    </property>
-  </bean>
-
-  <bean id="spnegoEntryPoint"
-    class="org.springframework.security.kerberos.web.SpnegoEntryPoint" >
-    <constructor-arg value="/login" />
-  </bean>
-
-  <bean id="spnegoAuthenticationProcessingFilter"
-    class="org.springframework.security.kerberos.web.SpnegoAuthenticationProcessingFilter">
-    <property name="authenticationManager" ref="authenticationManager" />
-  </bean>
-
-  <bean id="kerberosServiceAuthenticationProvider"
-    class="org.springframework.security.kerberos.KerberosServiceAuthenticationProvider">
-    <property name="ticketValidator">
-      <bean
-        class="org.springframework.security.kerberos.SunJaasKerberosTicketValidator">
-        <property name="servicePrincipal" value="${app.service-principal}" />
-        <property name="keyTabLocation" value="${app.keytab-location}" />
-        <property name="debug" value="true" />
-      </bean>
-    </property>
-    <property name="userDetailsService" ref="dummyUserDetailsService" />
-  </bean>
-
-  <bean id="dummyUserDetailsService"
-    class="org.springframework.security.kerberos.docs.DummyUserDetailsService" />
-
-</beans>
-<!-- end::snippetA[] -->
