+++ spring-security-kerberos-samples/sec-server-spnego-form-auth-xml/src/test/java/demo/app/ApplicationTest.java
@@ -1,0 +1,15 @@
+package demo.app;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringApplicationConfiguration(classes = Application.class)
+public class ApplicationTest {
+
+	@Test
+	public void configurationLoads() {
+	}
+}
